+++ src/index.ts
@@ -2,3 +2,4 @@ export { TangoClient } from "./client.js";
 export * from "./config.js";
 export * from "./errors.js";
 export * from "./types.js";
+export * from "./models/index.js";