+++ src/models/Location.ts
@@ -1,0 +1,16 @@
+export interface Location {
+  address_line1?: string | null;
+  address_line2?: string | null;
+  city?: string | null;
+  state?: string | null;
+  state_code?: string | null;
+  zip_code?: string | null;
+  zip?: string | null;
+  zip4?: string | null;
+  country?: string | null;
+  country_code?: string | null;
+  county?: string | null;
+  congressional_district?: string | null;
+  latitude?: number | null;
+  longitude?: number | null;
+}